refac: segregate models from entity
