+++ src/models/request/user/user-request.types.ts
@@ -1,0 +1,7 @@
+export interface UserRequest {
+    name: string;
+    email: string;
+    password: string;
+    birthDate: string;
+}
+
