+++ src/models/response/user/user-response.types.ts
@@ -1,0 +1,7 @@
+export interface UserResponse {
+    id: number;
+    name: string;
+    email: string;
+    birthDate: Date;
+}
+
