+++ src/repository/db-repository.ts
@@ -1,5 +1,6 @@
 import prisma from "../client/client";
-import { UserEntity, UserRequest } from "../entities/user-entity.types";
+import { UserEntity } from "../entities/user-entity.types";
+import { UserRequest } from "../models/request/user/user-request.types";
 
 
 export async function getUsers(): Promise<UserEntity[]> {
